:rewind: Revert stub for tests
